requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.6
django-crispy-forms==2.2
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
djangorestframework-3.15.2
